Basic refactoring of Calendar into routing component
