+++ app/javascript/pages/components/App.jsx
@@ -32,10 +32,10 @@ const App = (props) => (
         />
         <Route exact path="/gallery" component={Gallery} />
         <Route exact path="/login" component={Login} />
-        <Route path='/calendar/:year/:month' component={Calendar} />
-        <Route path='/browser/datasets/:id' component={DataViewer} />
-        <Route path='/browser/:menuOneSelectedItem' component={Browser} />
-        <Route path='/browser' component={Browser} />
+        <Route path="/calendar/:year/:month" component={Calendar} />
+        <Route path="/browser/datasets/:id" component={DataViewer} />
+        <Route path="/browser/:menuOneSelectedItem" component={Browser} />
+        <Route path="/browser" component={Browser} />
         <PrivateRoute exact path="/admin" component={Admin} />
       </Switch>
     </main>
